feat: 플레이리스트 제목/설명 수정 로직 추가
